refactor(ironic-understack): reorganize package into drivers/hooks/middleware/utils/conf
Groups the netdev hardware type and Redfish inspect interfaces under
drivers/, the Ironic inspection hooks under hooks/, the WSGI API
middleware under middleware/, shared helpers under utils/, and the
oslo_config setup under conf/, instead of a flat module layout.
Updates entry points in pyproject.toml and import paths accordingly.
