+++ python/ironic-understack/ironic_understack/hooks/inspect_hook_chassis_model.py
