+++ python/ironic-understack/ironic_understack/hooks/inspect_hook_node_name_check.py
