+++ python/ironic-understack/pyproject.toml
@@ -20,21 +20,21 @@ dependencies = [
 kubernetes = "ironic.console.container.kubernetes:KubernetesConsoleContainer"
 
 [project.entry-points."ironic.inspection.hooks"]
-resource-class = "ironic_understack.resource_class:ResourceClassHook"
-update-baremetal-port = "ironic_understack.inspect_hook_update_baremetal_ports:InspectHookUpdateBaremetalPorts"
-port-bios-name = "ironic_understack.port_bios_name_hook:PortBiosNameHook"
-node-name-check = "ironic_understack.inspect_hook_node_name_check:InspectHookNodeNameCheck"
-chassis_model = "ironic_understack.inspect_hook_chassis_model:InspectHookChassisModel"
+resource-class = "ironic_understack.hooks.resource_class:ResourceClassHook"
+update-baremetal-port = "ironic_understack.hooks.inspect_hook_update_baremetal_ports:InspectHookUpdateBaremetalPorts"
+port-bios-name = "ironic_understack.hooks.port_bios_name_hook:PortBiosNameHook"
+node-name-check = "ironic_understack.hooks.inspect_hook_node_name_check:InspectHookNodeNameCheck"
+chassis_model = "ironic_understack.hooks.inspect_hook_chassis_model:InspectHookChassisModel"
 
 [project.entry-points."ironic.hardware.types"]
-netdev = "ironic_understack.netdev_hardware:NetdevHardware"
+netdev = "ironic_understack.drivers.netdev_hardware:NetdevHardware"
 
 [project.entry-points."ironic.hardware.interfaces.inspect"]
-redfish-understack = "ironic_understack.redfish_inspect_understack:UnderstackRedfishInspect"
-idrac-redfish-understack = "ironic_understack.redfish_inspect_understack:UnderstackDracRedfishInspect"
+redfish-understack = "ironic_understack.drivers.redfish_inspect_understack:UnderstackRedfishInspect"
+idrac-redfish-understack = "ironic_understack.drivers.redfish_inspect_understack:UnderstackDracRedfishInspect"
 
 [project.entry-points."ironic.api.middleware"]
-portgroup-name-validation = "ironic_understack.portgroup_name_middleware:PortgroupNameValidationMiddleware"
+portgroup-name-validation = "ironic_understack.middleware.portgroup_name_middleware:PortgroupNameValidationMiddleware"
 
 [dependency-groups]
 test = [
